Ported the functionality of build.sh file to Makefile
The build.sh file was complicated and becomming more so,
thus now it's functionality of building the project and
deleting old files while compiling c++ programmes will
be done by Makefile.

The build.sh file was earlier called inside the main.sh script
This caused the execuation of main.sh script very slow.
Thus the main.sh script no longer builds the project, this should
be done by calling the "make" command in the machine only for the
first time.